improve ergonomics & split txs and txs convertion
