+++ data-service-consumer-rs/src/lib/utils.rs
@@ -17,6 +17,6 @@ pub fn epoch_ms_to_naivedatetime(ts: i64) -> NaiveDateTime {
     NaiveDateTime::from_timestamp(ts / 1000, (ts % 1000) as u32 * 1_000_000)
 }
 
-pub fn escape_unicode_null(s: &str) -> String {
-    s.replace("\0", "\\0")
+pub fn escape_unicode_null(s: impl AsRef<str>) -> String {
+    s.as_ref().replace("\0", "\\0")
 }
